+++ src/proof_system/widget/range/proverkey.rs
@@ -76,7 +76,7 @@ impl<F: PrimeField> ProverKey<F> {
 // Computes f(f-1)(f-2)(f-3)
 pub(crate) fn delta<F: PrimeField>(f: F) -> F {
     let f_1 = f - F::one();
-    let f_2 = f - F::from(2 as u64);
-    let f_3 = f - F::from(3 as u64);
+    let f_2 = f - F::from(2_u64);
+    let f_3 = f - F::from(3_u64);
     f * f_1 * f_2 * f_3
 }
